Overlay jdk with jdk11_headless and support overriding nixpkgs
When building the latest `master` with the pinned `nixpkgs` I noticed that the OpenJDK 8 derivation was being built. This PR overlays `jdk` to point to `jdk11_headless` so that it is also used by `maven`. Seems like a good idea to me to keep dependency count low but maybe there's a more idiomatic way to achieve this? In addition, I made it possible to bring one's own `nixpkgs` if one is so inclined -- I certainly am :smile: 


* Overlay jdk with jdk11_headless
* Support bringing one's own nixpkgs